fix check_balance

+++ internal/models/withdraw/balance.go
@@ -1,6 +1,6 @@
 package withdraw
 
 type Balance struct {
-	Current   int64 `json:"current"`
-	Withdrawn int64 `json:"withdrawn"`
+	Current   float32 `json:"current"`
+	Withdrawn float32 `json:"withdrawn"`
 }
