+++ internal/models/withdraw/withdraw.go
@@ -5,6 +5,6 @@ import "time"
 type Withdraw struct {
 	User        string     `json:"-"`
 	Order       string     `json:"order"`
-	Sum         int64      `json:"sum"`
+	Sum         float32    `json:"sum"`
 	ProcessedAt *time.Time `json:"processed_at"`
 }
